+++ frontend/packages/erd-core/src/features/erd/components/ERDRenderer/LeftPane/LeftPane.tsx
@@ -19,9 +19,9 @@ import { useCallback, useMemo } from 'react'
 import { useVersionOrThrow } from '../../../../../providers'
 import { useUserEditingOrThrow } from '../../../../../stores'
 import { useCustomReactflow } from '../../../../reactflow/hooks'
-import { useTableVisibility } from '../../../hooks/useTableVisibility/useTableVisibility'
 import { isTableNode } from '../../../utils'
 import { updateNodesHiddenState } from '../../ERDContent/utils'
+import { useTableVisibility } from '../hooks'
 import { CopyLinkButton } from './CopyLinkButton'
 import styles from './LeftPane.module.css'
 import { MenuItemLink, type Props as MenuItemLinkProps } from './MenuItemLink'
