+++ frontend/packages/erd-core/src/features/erd/components/ERDRenderer/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from './useTableVisibility'
