+++ frontend/packages/erd-core/src/features/erd/components/ERDRenderer/hooks/useTableVisibility/index.ts
@@ -1,0 +1,1 @@
+export * from './useTableVisibility'
